requirements.txt
streamlit
google-generativeai
streamlit-chat
Pillow